moved buffer to same dir
